Send library click to new event handler
Instead of the library click only opening the modal window, now there is
a new event handler to handle clicks within the library element.
Somteimes a library click is intended to open a modal and sometimes it
is intended to remove a book from the library. The book removal logic
doesn't exist yet.