Refactor scripts so that actions are at beginning
The add book button will now be at the beginning of the library section
in the DOM so that it doesn't get pushed down the page. Users will have
to scroll to view books regardless, but I don't want them to have to
scroll in search of the action button. In tandem with that, any place I
display myLibrary, I'm reversing the order of the array before displaying
so that the most recent additions appear next to the add-book buttin --
where the action is taking place.